(refactor) adblock detection
