+++ app/src/hooks/index.ts
@@ -1,15 +1,16 @@
+export * from './connectedWeb3'
+export * from './geoJs'
+export { use24hsVolume } from './use24hsVolume'
 export { useAsyncDerivedValue } from './useAsyncDerivedValue'
+export { useBlockchainMarketMakerData } from './useBlockchainMarketMakerData'
 export { useCheckContractExists } from './useCheckContractExists'
 export { useCollateral } from './useCollateral'
 export { useCollateralBalance } from './useCollateralBalance'
-export * from './connectedWeb3'
-export * from './geoJs'
 export { useContracts } from './useContracts'
+export { useDetectAdblocker } from './useDetectAdblocker'
+export { useDocumentDescription } from './useDocumentDescription'
+export { useDocumentTitle } from './useDocumentTitle'
 export { useFundingBalance } from './useFundingBalance'
-export { useBlockchainMarketMakerData } from './useBlockchainMarketMakerData'
 export { useGraphMarketMakerData } from './useGraphMarketMakerData'
 export { useMarketMakerData } from './useMarketMakerData'
 export { usePolling } from './usePolling'
-export { useDocumentTitle } from './useDocumentTitle'
-export { useDocumentDescription } from './useDocumentDescription'
-export { use24hsVolume } from './use24hsVolume'
